Implementing 360 Degree Videos
There are two ways to ingest 360 videos in an Instant Article. The first is by uploading them, where you point directly to the media file and use the native Instant Articles video player. The other is to embed them, where you simply embed the 360 video you've already uploaded to Facebook.